+++ lib/view_component/slotable.rb
@@ -254,10 +254,7 @@ module ViewComponent
 
       def validate_plural_slot_name(slot_name)
         if RESERVED_NAMES[:plural].include?(slot_name.to_sym)
-          raise ArgumentError.new(
-            "#{self} declares a slot named #{slot_name}, which is a reserved word in the ViewComponent framework.\n\n" \
-            "To fix this issue, choose a different name."
-          )
+          raise ReservedPluralSlotNameError.new(self.name, slot_name)
         end
 
         raise_if_slot_ends_with_question_mark(slot_name)